[Refactor] : AddScheduleView 이름 변경
Git : #1
