+++ lib/presentation/addschedule/view/appbar.dart
@@ -1,4 +1,4 @@
-part of 'addschedule_view.dart';
+part of 'view.dart';
 
 class _AppBar extends AppBar {
   BuildContext context;
